Revert "chore: Run cargo hack on each workspace member and all targets (#11419)"
This reverts commit 3596e310e6abe27a70c5d88deaf6d90c8c94637a.
