+++ lib/tracing-limit/Cargo.toml
@@ -9,14 +9,14 @@ license = "MPL-2.0"
 [dependencies]
 ansi_term = { version = "0.12", default-features = false }
 tracing-core = { version = "0.1", default-features = false }
-tracing-subscriber = { version = "0.3", default-features = false, features = ["registry", "std"] }
+tracing-subscriber = { version = "0.3", default-features = false, features = ["std"] }
 dashmap = { version = "5", default-features = false }
 
 [dev-dependencies]
 criterion = "0.3"
 tracing = "0.1.30"
 mock_instant = { version = "0.2" }
-tracing-subscriber = { version = "0.3.8", default-features = false, features = ["env-filter", "fmt"] }
+tracing-subscriber = { version = "0.3.8", default-features = false, features = ["env-filter"] }
 
 [[bench]]
 name = "limit"
