+++ lib/tracing-limit/examples/by_span.rs
@@ -1,4 +1,7 @@
-use tracing::{info, info_span, trace, Dispatch};
+#[macro_use]
+extern crate tracing;
+
+use tracing::Dispatch;
 use tracing_limit::RateLimitedLayer;
 use tracing_subscriber::layer::SubscriberExt;
 
@@ -12,7 +15,7 @@ fn main() {
     let dispatch = Dispatch::new(subscriber);
 
     tracing::dispatcher::with_default(&dispatch, || {
-        for i in 0..40usize {
+        for i in 0..40 {
             trace!("This field is not rate limited!");
             for key in &["foo", "bar"] {
                 for line_number in &[1, 2] {
