+++ lib/vector-buffers/Cargo.toml
@@ -27,7 +27,7 @@ tokio-stream = { version = "0.1", default-features = false, features = ["sync"] 
 tokio-util = { version = "0.6", default-features = false }
 tokio = { version = "1.16.1", default-features = false, features = ["rt", "macros", "rt-multi-thread", "sync", "fs", "io-util", "time"] }
 tracing = { version = "0.1.30", default-features = false, features = ["attributes"] }
-vector_common = { path = "../vector-common", default-features = false, features = ["byte_size_of"] }
+vector_common = { path = "../vector-common", default-features = false }
 
 [dev-dependencies]
 clap = "3.1.0"
