+++ lib/vector-common/Cargo.toml
@@ -21,8 +21,6 @@ aws_cloudwatch_logs_subscription = [
 
 btreemap = []
 
-byte_size_of = ["bytes", "chrono"]
-
 conversion = [
   "bytes",
   "chrono",
