+++ lib/vector-common/src/lib.rs
@@ -12,7 +12,6 @@ pub mod aws_cloudwatch_logs_subscription;
 #[cfg(feature = "btreemap")]
 pub mod btreemap;
 
-#[cfg(feature = "byte_size_of")]
 pub mod byte_size_of;
 
 #[cfg(feature = "conversion")]
