+++ lib/vrl/stdlib/Cargo.toml
@@ -43,7 +43,7 @@ roxmltree = { version = "0.14.1", optional = true }
 woothee = { version = "0.13.0", optional = true }
 uaparser = { version = "0.5.0", optional = true }
 utf8-width = { version = "0.1.5", optional = true }
-vector_common = { path = "../../vector-common", default-features = false, features = ["btreemap"] }
+vector_common = { path = "../../vector-common", default-features = false, optional = true }
 
 [dev-dependencies]
 anyhow = "1"
@@ -188,13 +188,13 @@ decode_percent = ["percent-encoding"]
 del = []
 downcase = []
 encode_base64 = ["base64"]
-encode_json = ["serde_json", "value/json", "chrono", "regex"]
+encode_json = ["serde_json", "value/json"]
 encode_key_value = ["vector_common/encoding", "value/json"]
 encode_logfmt = ["encode_key_value"]
 encode_percent = ["percent-encoding"]
 ends_with = []
 exists = []
-find = ["regex"]
+find = []
 find_table_row = []
 flatten = []
 float = []
@@ -220,9 +220,9 @@ is_integer = []
 is_null = []
 is_nullish = []
 is_object = []
-is_regex = ["regex"]
+is_regex = []
 is_string = []
-is_timestamp = ["chrono"]
+is_timestamp = []
 join = []
 length = []
 log = ["tracing"]
@@ -236,7 +236,7 @@ now = ["chrono"]
 object = []
 parse_apache_log = ["chrono", "once_cell", "regex", "vector_common/conversion"]
 parse_aws_alb_log = ["nom"]
-parse_aws_cloudwatch_log_subscription_message = ["serde_json", "vector_common/aws_cloudwatch_logs_subscription", "vector_common/btreemap", "chrono"]
+parse_aws_cloudwatch_log_subscription_message = ["serde_json", "vector_common/aws_cloudwatch_logs_subscription", "vector_common/btreemap"]
 parse_aws_vpc_flow_log = []
 parse_common_log = ["chrono", "once_cell", "regex", "vector_common/conversion"]
 parse_csv = ["csv"]
@@ -256,7 +256,7 @@ parse_regex = ["regex"]
 parse_regex_all = ["regex"]
 parse_ruby_hash = ["nom"]
 parse_syslog = ["syslog_loose", "chrono", "vector_common/conversion"]
-parse_timestamp = ["vector_common/conversion", "chrono"]
+parse_timestamp = ["vector_common/conversion"]
 parse_tokens = ["vector_common/tokenize"]
 parse_url = ["url"]
 parse_user_agent = ["woothee","uaparser","once_cell"]
@@ -264,7 +264,7 @@ parse_xml = ["roxmltree", "once_cell", "regex"]
 push = []
 redact = ["once_cell", "regex"]
 remove = ["vector_common/btreemap"]
-replace = ["regex"]
+replace = []
 reverse_dns = ["dns-lookup"]
 round = []
 set = ["vector_common/btreemap"]
@@ -272,18 +272,18 @@ sha1 = ["sha-1", "hex"]
 sha2 = ["sha-2", "hex"]
 sha3 = ["sha-3", "hex"]
 slice = []
-split = ["regex"]
+split = []
 starts_with = ["utf8-width"]
 string = []
 strip_ansi_escape_codes = ["bytes", "strip-ansi-escapes"]
 strip_whitespace = []
-tag_types_externally = ["vector_common/btreemap", "chrono", "regex"]
+tag_types_externally = ["vector_common/btreemap"]
 tally = []
 tally_value = []
 timestamp = []
 to_bool = ["vector_common/conversion"]
-to_float = ["vector_common/conversion", "chrono"]
-to_int = ["vector_common/conversion", "chrono"]
+to_float = ["vector_common/conversion"]
+to_int = ["vector_common/conversion"]
 to_regex = ["tracing", "regex"]
 to_string = ["chrono"]
 to_syslog_facility = []
@@ -304,4 +304,3 @@ bench = false
 name = "benches"
 harness = false
 test = true
-required-features = ["default"]
