+++ src/conditions/mod.rs
@@ -20,10 +20,6 @@ pub enum Condition {
     Vrl(vrl::Vrl),
     CheckFields(check_fields::CheckFields),
     DatadogSearch(datadog_search::DatadogSearchRunner),
-
-    // used for benchmarks
-    AlwaysPass,
-    AlwaysFail,
 }
 
 impl Condition {
@@ -35,8 +31,6 @@ impl Condition {
             Condition::CheckFields(x) => x.check(e),
             Condition::DatadogSearch(x) => x.check(e),
             Condition::Vrl(x) => x.check(e),
-            Condition::AlwaysPass => true,
-            Condition::AlwaysFail => false,
         }
     }
 
@@ -50,8 +44,6 @@ impl Condition {
             Condition::CheckFields(x) => x.check_with_context(e),
             Condition::DatadogSearch(x) => x.check_with_context(e),
             Condition::Vrl(x) => x.check_with_context(e),
-            Condition::AlwaysPass => Ok(()),
-            Condition::AlwaysFail => Ok(()),
         }
     }
 }
